# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The official UHRP Storage Server implementation in Go, allowing anyone to host their own bitcom-based public file CDN.",
        "title": "UHRP Storage Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/find": {
            "get": {
                "description": "Find metadata (size, mimeType, expiry) for a specific UHRP file advertisement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Find file metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UHRP URL of the file to find",
                        "name": "uhrpUrl",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.findResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "Retrieve a list of all currently active (non-expired) UHRP advertisements created by the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List user's active file uploads",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.listResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/put": {
            "put": {
                "description": "Upload a file's binary content using a pre-signed URL generated by the /upload endpoint.",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Uploader's identity key",
                        "name": "uploader",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique object identifier",
                        "name": "objectID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Expected file size in bytes",
                        "name": "fileSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC expiry timestamp",
                        "name": "expiry",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC signature validating the upload request",
                        "name": "hmac",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quote": {
            "post": {
                "description": "Get a price quote in satoshis for hosting a file of a specific size for a specific duration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Calculate storage price",
                "parameters": [
                    {
                        "description": "File size and retention period details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.quoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.quoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/renew": {
            "post": {
                "description": "Extend the storage time for an existing UHRP file advertisement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Renew an active file",
                "parameters": [
                    {
                        "description": "UHRP URL and additional minutes to host",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.renewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.renewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Get a pre-signed URL and payment details to upload a file to the UHRP storage server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Request an upload URL",
                "parameters": [
                    {
                        "description": "File size and desired retention period",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.uploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.uploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.findData": {
            "type": "object",
            "properties": {
                "expiryTime": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                }
            }
        },
        "handlers.findResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/handlers.findData"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.listResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uploads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.listUpload"
                    }
                }
            }
        },
        "handlers.listUpload": {
            "type": "object",
            "properties": {
                "expiryTime": {
                    "type": "integer"
                },
                "uhrpUrl": {
                    "type": "string"
                }
            }
        },
        "handlers.quoteRequest": {
            "type": "object",
            "properties": {
                "fileSize": {
                    "type": "integer",
                    "example": 1024
                },
                "retentionPeriod": {
                    "type": "integer",
                    "example": 60
                }
            }
        },
        "handlers.quoteResponse": {
            "type": "object",
            "properties": {
                "quote": {
                    "type": "integer"
                }
            }
        },
        "handlers.renewRequest": {
            "type": "object",
            "properties": {
                "additionalMinutes": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "uhrpUrl": {
                    "type": "string"
                }
            }
        },
        "handlers.renewResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "newExpiryTime": {
                    "type": "integer"
                },
                "prevExpiryTime": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.uploadRequest": {
            "type": "object",
            "properties": {
                "fileSize": {
                    "type": "integer"
                },
                "retentionPeriod": {
                    "type": "integer"
                }
            }
        },
        "handlers.uploadResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "requiredHeaders": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "uploadURL": {
                    "type": "string"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthfetchIdentity": {
            "description": "Authentication via go-bsv-middleware using BRC-43 Authfetch.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}